% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_panel_information.R
\name{get_panel_information}
\alias{get_panel_information}
\title{Provides Axiom Human Origins panel ascertainment information}
\usage{
get_panel_information(panel = "all")
}
\arguments{
\item{panel}{A list of panels for which ascertainment information is desired}
}
\value{
a table of ascertainment information
}
\description{
\code{get_panel_information} returns ascertainment information for the various SNP
panels.
}
