% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_axiom_snps.R
\name{get_axiom_snps}
\alias{get_axiom_snps}
\title{Get a list of SNP RS ids from desired Axiom Human Origins ascertainment Panels}
\usage{
get_axiom_snps(panel)
}
\arguments{
\item{panel}{a list of panels to be included.}
}
\value{
a list of SNP RS ids.
}
\description{
\code{get_axiom_snps} returns a list of snps corresponding to requested Human Origins ascertainment panels
}
