% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sample_information.R
\name{get_sample_information}
\alias{get_sample_information}
\title{Gets sample information for a list of sample IDs}
\usage{
get_sample_information(ID, na.fill = TRUE)
}
\arguments{
\item{ID}{a list of sample IDs}

\item{na.fill}{a boolean indicating whether to fill missing data with NA
(default TRUE)}
}
\value{
a data frame of sample id, population, and region
}
\description{
\code{get_sample_information} takes a list of sample IDs and returns population
and region information for each sample
}
